refactor DefaultTrace
