+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/RandomUuidGenerator.java
@@ -1,0 +1,14 @@
+package org.springframework.cloud.sleuth;
+
+import java.util.UUID;
+
+/**
+ * @author Spencer Gibb
+ */
+public class RandomUuidGenerator implements IdGenerator {
+
+	@Override
+	public String create() {
+		return UUID.randomUUID().toString();
+	}
+}
