+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/Trace.java
@@ -81,7 +81,5 @@ public interface Trace {
 	 */
 	boolean isTracing();
 
-	public void deliver(Span span);
-
-	void setSpanReceivers(Collection<SpanReceiver> spanReceivers);
+	void deliver(Span span);
 }
