+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/sampler/IsTracingSampler.java
@@ -1,20 +1,15 @@
 package org.springframework.cloud.sleuth.trace.sampler;
 
 import org.springframework.cloud.sleuth.trace.Sampler;
-import org.springframework.cloud.sleuth.trace.Trace;
+import org.springframework.cloud.sleuth.trace.SpanHolder;
 
 /**
  * @author Spencer Gibb
  */
 public class IsTracingSampler implements Sampler<Object> {
-	private final Trace trace;
-
-	public IsTracingSampler(Trace trace) {
-		this.trace = trace;
-	}
 
 	@Override
 	public boolean next(Object info) {
-		return this.trace.isTracing();
+		return SpanHolder.getCurrentSpan() != null;
 	}
 }
